requirements.txt
Django
django-import-export
